+++ services/backend/src/main/java/com/soprasteria/g4it/backend/apiadministratoractions/controller/AdministratorActionsRestController.java
@@ -1,7 +1,7 @@
 package com.soprasteria.g4it.backend.apiadministratoractions.controller;
 
 import com.soprasteria.g4it.backend.apiadministratoractions.business.AdministratorActionsService;
-import com.soprasteria.g4it.backend.apiadministratoractions.business.RoleManagementService;
+import com.soprasteria.g4it.backend.apiadministratoractions.business.DsMigrationService;
 import com.soprasteria.g4it.backend.server.gen.api.AdministratorActionsApiDelegate;
 import com.soprasteria.g4it.backend.server.gen.api.dto.AllEvaluationStatusRest;
 import lombok.extern.slf4j.Slf4j;
@@ -14,10 +14,26 @@ import org.springframework.stereotype.Service;
 public class AdministratorActionsRestController implements AdministratorActionsApiDelegate {
     @Autowired
     AdministratorActionsService administratorActionsService;
+    @Autowired
+    DsMigrationService dsMigrationService;
 
-
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public ResponseEntity<AllEvaluationStatusRest> doAdminActions() {
+
+        // Migration of digital services from DEMO workspace to new workspaces
+        return ResponseEntity.ok(dsMigrationService.migrateDemoDs());
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public ResponseEntity<AllEvaluationStatusRest> removeWriteAccess() {
+
+      //  Remove write access for digital services on demo organizations
         return ResponseEntity.ok(administratorActionsService.executeWriteRoleCleanupOnDemoOrg());
     }
 }